refactor: move symbol parser in Note
In Note we need a reference to the symbol that generated it so that
we can easily convert Tunes back to their textual representation
without further computation.

If we left the open contructor in Note, we would risk the case where
Notes could be constructed without a Symbol, breaking the convention
above.

To solve the issues, we moved the parser into Note::fromSymbol and
promoted it to be the only publicly available constructor and prevent
the problem from happening altogether.
