+++ latebit/sid/synth/track.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <string>
 #include <vector>
 
 #include "oscillator.h"
@@ -7,41 +8,56 @@
 using namespace std;
 
 namespace sid {
-// A Note is a musical symbol that represents a sound. It has a pitch, a volume,
+// Symbol is a string that represents a Note. It can be a rest, a continue,
+// the end of a track or a note. It is used to parse and generate notes.
+using Symbol = string;
+
+const string REST_SYMBOL = "------";
+const string CONTINUE_SYMBOL = "......";
+const string END_OF_TRACK_SYMBOL = "      ";
+
+// A Note is a musical object that represents a sound. It has a pitch, a volume,
 // a wave type and an effect type. It also has a unique identifier that can be
-// used to compare notes.
+// used to compare notes. Notes are immutable, which means that once they are
+// created, they cannot be modified.
 // A note can also be a rest, which is a symbol that represents silence.
 // A note can also be invalid, which is a symbol that represents an invalid note
 // and it is used to report errors.
 class Note {
  public:
-  Note(int pitch, int volume, WaveType wave, EffectType effect);
   // Creates a new rest note
-  static auto rest() -> Note;
+  static auto makeRest() -> Note;
   // Creates a new invalid note
-  static auto invalid() -> Note;
+  static auto makeInvalid() -> Note;
+  // Creates a new continue note
+  static auto makeContinue() -> Note;
+  // Creates a new note from a symbol
+  static auto fromSymbol(const Symbol& str) -> Note;
 
   // Checks if the given note is a rest
-  auto isRest() -> bool;
-  // Checks if the given notes are the same (i.e., have the same id)
-  auto isSame(Note other) -> bool;
+  [[nodiscard]] auto isRest() const -> bool;
   // Checks if the given notes are equal (i.e., same pitch, volume, wave)
-  auto isEqual(Note other) -> bool;
+  [[nodiscard]] auto isEqual(Note other) const -> bool;
   // Checks if the given note is invalid
-  auto isInvalid() -> bool;
+  [[nodiscard]] auto isInvalid() const -> bool;
+  // Checks if the given note is continue
+  [[nodiscard]] auto isContinue() const -> bool;
 
   // Returns the pitch of the note
-  auto getPitch() -> int;
+  [[nodiscard]] auto getPitch() const -> int;
   // Returns the volume of the note in the range [0, 1]
-  auto getVolume() -> float;
+  [[nodiscard]] auto getVolume() const -> float;
   // Returns the wave type of the note
-  auto getWave() -> WaveType;
+  [[nodiscard]] auto getWave() const -> WaveType;
   // Returns the effect type of the note
-  auto getEffect() -> EffectType;
+  [[nodiscard]] auto getEffect() const -> EffectType;
   // Returns the unique identifier of the note
-  auto getId() -> long unsigned int;
+  [[nodiscard]] auto getId() const -> long unsigned int;
+  // Returns the symbol from which this note was created
+  [[nodiscard]] auto getSymbol() const -> Symbol;
 
  private:
+  Note(int pitch, int volume, WaveType wave, EffectType effect, string symbol);
   // A number between 0 and 96 describing a note from C0 to B7
   int pitch = 0;
   // A number between 0 and 15 describing the volume of the note
@@ -51,7 +67,9 @@ class Note {
   // The type of effect associated with the note
   EffectType effect = NONE;
   // The unique identifier of the note, used to compare notes
-  long unsigned int id = 0;
+  long int id = 0;
+  // The symbol associated with the note
+  Symbol symbol = END_OF_TRACK_SYMBOL;
 };
 
 using Track = std::vector<Note>;
